Source bash_profile in pre_commit hook
Fork (the git client) doesn't seem to source zshrc, bashrc,
bash_profile or profile in whatever shell it uses to run git commands.
Manually sourcing bash_profile, if it exists, makes sure that $GOPATH
is available to the hooks. This, of course, means that $GOPATH must
be exported from ~/.bash_profile but I can't find a better solution.