Remove the service. prefix from service names
Since Kubernetes DNS names can't have dots in them, the k8s service
names ended up being transformed from service.foo to s-foo. This meant
the hostname of services was different in prod (k8s) vs dev (Docker).
Getting around this problem is a pain because it doesn't seem possible
to set a DNS alias for a k8s service, and adding an alias to each
service in the docker-compose file is quite verbose.

Instead, the problem can be avoided altogether by removing dots from the
service names. There is no need for the whole service. prefix. This PR
doesn't change the directory names or structure, but they could be moved
to a services/ folder and have the prefix removed from the directories
if desired.