Don't make `any` types pointers
The `any` type in JRPC becomes an interface{} in go, which is already a
reference type, so making it a pointer is an unnecessary level of
redirection. This change treats it like the other reference types (maps
and slices) which also don't need to be explicit pointers to determine
whether the value was set in the JSON (once the JSON is unmarshaled).