Don't re-panic after recovers in dev
This was originally done so that the stack trace is pretty-printed in
the terminal. When printed via slog, the new line and tab characters are
printed in their escaped format. However, the panic prevents the
response from being written to the client. It also means the behaviour
is different between dev and prod. This problem should be resolved in a
different way.